Machine Learning/ASL Hand Gesture/WIP Deployment/requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy